Reduce scope of `rescue` in #hide_original_method
Having temporarily removed the `rescue` statement, I can see that we're
trying to catch `NameError`s that are raised when we try to store the
original method. We only have tests to ensure that we're catching these
exceptions in AnyInstanceMethod but I figure it's probably safest to
retain the behaviour in ClassMethod too.

The changes in this commit are easier to see/understand if you use the
`--ignore-all-space` option.